processing/requirements.txt
apache-beam==2.3.0
avro==1.8.2
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
crcmod==1.7
dill==0.2.6
enum34==1.1.6
fasteners==0.14.1
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gapic-google-cloud-pubsub-v1==0.15.4
google-api-core==1.1.0
google-apitools==0.5.20
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==0.25.0
google-cloud-core==0.25.0
google-cloud-language==1.0.1
google-cloud-pubsub==0.26.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
googledatastore==7.0.1
grpc-google-iam-v1==0.11.4
grpcio==1.10.0
hdfs3==0.3.0
httplib2==0.9.2
idna==2.6
mock==2.0.0
monotonic==1.4
oauth2client==4.1.2
pbr==3.1.1
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2018.3
PyVCF==0.6.8
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
typing==3.6.4
urllib3==1.22
